requirements.txt
# Core DB clients
psycopg2-binary>=2.9.9
psycopg[binary]>=3.2.0
psycopg-pool>=3.2.0

# AI / API
openai>=1.50.0
requests>=2.32.0
python-dotenv>=1.0.1

# Data / validation / math
jsonschema>=4.23.0
sympy>=1.13.3
Pillow>=10.4.0
PyMuPDF>=1.24.10

# Test dependencies used by Python test suites
pytest>=9.0.0
hypothesis>=6.100.0
